% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_bc.R
\name{calc_bc}
\alias{calc_bc}
\title{Calculate Bray-Curtis dissimilarity}
\usage{
calc_bc(
  df_collapse_mock,
  df_collapse_vpf,
  df_collapse_brackenVPF,
  df_collapse_brackenSTD,
  vlevel
)
}
\arguments{
\item{df_collapse_mock}{Abundance profile of mock taxonomy}

\item{df_collapse_vpf}{Abundance profile of ViroProfiler taxonomy}

\item{df_collapse_brackenVPF}{Abundance profile of Bracken on ViroProfiler taxonomy}

\item{df_collapse_brackenSTD}{Abundance profile of Bracken on kraken2 STD taxonomy}

\item{vlevel}{Taxonomy level}
}
\value{
DataFrame
}
\description{
Calculate Bray-Curtis dissimilarity
}
