% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_confmat.R
\name{calc_confmat}
\alias{calc_confmat}
\title{Calculate confusion matrix}
\usage{
calc_confmat(
  df_collapse_mock,
  df_collapse_vpf,
  df_collapse_brackenVPF,
  df_collapse_brackenSTD,
  vlevel,
  minrel
)
}
\arguments{
\item{df_collapse_mock}{Abundance profile of mock taxonomy}

\item{df_collapse_vpf}{Abundance profile of ViroProfiler taxonomy}

\item{df_collapse_brackenVPF}{Abundance profile of Bracken on ViroProfiler taxonomy}

\item{df_collapse_brackenSTD}{Abundance profile of Bracken on kraken2 STD taxonomy}

\item{vlevel}{Taxonomy level}

\item{minrel}{Min relative abundance}
}
\value{
DataFrame
}
\description{
Calculate confusion matrix
}
