% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ab_bracken.R
\name{create_ab_bracken}
\alias{create_ab_bracken}
\title{Read Bracken abundance report file}
\usage{
create_ab_bracken(
  fin_ab_bracken,
  t2lin,
  cnames,
  vlevel,
  remove_unclassified = 0
)
}
\arguments{
\item{fin_ab_bracken}{Bracken abundance report file}

\item{t2lin}{Mapping file of taxonomy ID <-> lineage}

\item{cnames}{Selected column names}

\item{vlevel}{Taxonomy level}

\item{remove_unclassified}{Merge unclassified taxa or not, default 0}
}
\value{
DataFrame
}
\description{
Read Bracken abundance report file
}
