% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ab_bracken.R
\name{create_ab_brackenVPF}
\alias{create_ab_brackenVPF}
\title{Abundance profile of Bracken using ViroProfiler taxonomy}
\usage{
create_ab_brackenVPF(
  df_vpf_taxa,
  fin_brackenVPF,
  cnames,
  vlevel,
  remove_unclassified = 0
)
}
\arguments{
\item{df_vpf_taxa}{dataframe of ViroProfiler taxonomy}

\item{fin_brackenVPF}{file of ViroProfiler bracken abundance table}

\item{cnames}{Selected column names}

\item{vlevel}{taxonomy level}

\item{remove_unclassified}{Merge unclassified taxa or not, default 0}
}
\value{
DataFrame
}
\description{
Abundance profile of Bracken using ViroProfiler taxonomy
}
