% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ab_mock.R
\name{create_ab_mock}
\alias{create_ab_mock}
\title{Create taxonomy abundance of of mock data}
\usage{
create_ab_mock(fin_t2lin_mock, fins_ab_mock, vlevel, remove_unclassified = 0)
}
\arguments{
\item{fin_t2lin_mock}{mapping file of gi <-> taxid <-> lineage}

\item{fins_ab_mock}{abundance files of mock data}

\item{vlevel}{taxonomy level}

\item{remove_unclassified}{default: 0}
}
\value{
A dataframe
}
\description{
Create taxonomy abundance of of mock data
}
