% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ab_vpf.R
\name{create_ab_vpf}
\alias{create_ab_vpf}
\title{Abundance profile of ViroProfiler}
\usage{
create_ab_vpf(
  df_vpf_taxa,
  fin_vpf_count,
  fin_vpf_cov,
  cnames,
  vlevel,
  mincov = 0.1,
  bin2contig = 0,
  remove_unclassified = 0,
  fctglen = 0
)
}
\arguments{
\item{df_vpf_taxa}{dataframe of ViroProfiler taxonomy}

\item{fin_vpf_count}{file of ViroProfiler count table}

\item{fin_vpf_cov}{file of ViroProfiler coverage fraction table}

\item{cnames}{Selected column names}

\item{vlevel}{taxonomy level}

\item{mincov}{ViroProfiler minimum coverage fraction 0-1}

\item{bin2contig}{Mapping file of bins and contigs, default 0}

\item{remove_unclassified}{Merge unclassified taxa or not, default 0}

\item{fctglen}{Mapping file of contig length}
}
\value{
DataFrame
}
\description{
Abundance profile of ViroProfiler
}
