% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_prab.R
\name{create_prab}
\alias{create_prab}
\title{Convert abundance table to prab (present/absent) table}
\usage{
create_prab(df_abundance, df_taxa, minrel, vlevel, method_name)
}
\arguments{
\item{df_abundance}{Abundance table}

\item{df_taxa}{Table with all taxonomy name}

\item{minrel}{Min relative abundance}

\item{vlevel}{Taxonomy level}

\item{method_name}{Method name, Mock, ViroProfiler, etc}
}
\value{
DataFrame
}
\description{
Convert abundance table to prab (present/absent) table
}
