% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mock.R
\name{read_mock}
\alias{read_mock}
\title{Read mock abundance file}
\usage{
read_mock(fpath)
}
\arguments{
\item{fpath}{PATH that contains *_comp.tsv files}
}
\value{
A dataframe hello
}
\description{
Read mock abundance file
}
