% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vpf_prab}
\alias{vpf_prab}
\title{Taxonomy abundance profile in PRAB (present/absent) format}
\format{
A data frame with 7,240 rows and 60 columns:
\describe{
  \item{method}{ViroProfiler, BrackenVPF, BrackenSTD, Mock}
  \item{metric}{precision, recall, f1_score}
  \item{sample_id}{Sample ID}
  \item{taxalevel}{Taxonomy level}
  \item{threshold}{Threshold of relative abundance}
  \item{value}{Value of precision, recall, f1_score}
  ...
}
}
\source{
{Created in-house PAB profile}
data(vpf_prab)
}
\usage{
vpf_prab
}
\description{
Abundance profile of a specific taxonomy level
Report ...
}
\keyword{datasets}
