% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vpf_relab}
\alias{vpf_relab}
\title{Taxonomy abundance profile in RELAB (Relative abundance) format}
\format{
A data frame with 7,240 rows and 60 columns:
\describe{
  \item{bcs}{Bray-Curtis dissimilarity}
  \item{method}{ViroProfiler, BrackenVPF, BrackenSTD}
  \item{sample_id}{Sample ID}
  \item{taxalevel}{Taxonomy level}
  ...
}
}
\source{
{Created in-house PAB profile}
data(vpf_relab)
}
\usage{
vpf_relab
}
\description{
Abundance profile of a specific taxonomy level
Report ...
}
\keyword{datasets}
