% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpfpaper-package.R
\docType{package}
\name{vpfpaper-package}
\alias{vpfpaper}
\alias{vpfpaper-package}
\title{vpfpaper: ViroProfiler paper source code}
\description{
Source code, analyses of ViroProfiler paper
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rujinlong/vpfpaper}
  \item Report bugs at \url{https://github.com/rujinlong/vpfpaper/issues}
}

}
\author{
\strong{Maintainer}: Jinlong Ru \email{jinlong.ru@helmholtz-muenchen.de} (\href{https://orcid.org/0000-0002-6757-6018}{ORCID})

}
\keyword{internal}
